feat: rename img-commment to img-content
