+++ src/components/comments/comment/anniversaries/anniversaries.html
@@ -13,7 +13,7 @@
         </ion-item>
       </ion-list>
     </ion-row>
-    <img-comment *ngIf="comment.data.media" [src]=comment.data.media[0].url></img-comment>
+    <img-content *ngIf="comment.data.media" [src]=comment.data.media[0].url></img-content>
     <ion-row>
       <ion-col col-6></ion-col>
       <ion-col col-6 center text-center>
