+++ src/components/comments/comments.html
@@ -6,7 +6,7 @@
     </ion-avatar>    
     <h2></h2>
     <p></p>
-    <div class=img-comment>&nbsp;</div>
+    <div class=img-content>&nbsp;</div>
   </div>
 </ion-list>
 <ion-infinite-scroll *ngIf="!finished && comments" (ionInfinite)="loadMore($event)">
