+++ src/components/comments/comments.scss
@@ -11,7 +11,7 @@ comments {
         margin-left: 0px !important;
       }
     
-      .img-comment {
+      .img-content {
         max-height: 10em;
       }    
     
@@ -23,7 +23,7 @@ comments {
           margin-top: 10px;
         }
       
-        .img-comment {
+        .img-content {
           background-color: lightgrey;
           opacity: 0.3;
           width: 50%;
