+++ src/components/images/images.module.ts
@@ -1,12 +1,12 @@
 import { NgModule } from '@angular/core';
 import { ImgAvatarComponent } from './img-avatar/img-avatar';
 import { IonicModule } from 'ionic-angular';
-import { ImgCommentComponent } from '@components/images/img-comment/img-comment';
+import { ImgContentComponent } from '@components/images/img-content/img-content';
 import { SharedModule } from '@shared/shared.module';
 @NgModule({
 	declarations: [
 		ImgAvatarComponent,
-		ImgCommentComponent
+		ImgContentComponent
 	],
 	imports: [
 		IonicModule,
@@ -15,7 +15,7 @@ import { SharedModule } from '@shared/shared.module';
 	exports: [
 		IonicModule,
 		ImgAvatarComponent,
-		ImgCommentComponent
+		ImgContentComponent
 	]
 })
 export class ImagesModule {}
