--- src/components/images/img-comment/img-comment.html
@@ -1,1 +1,0 @@
-<img *ngIf="srcRetail" [src]="srcRetail" [srcset]="srcSet" imageViewer="{{srcZoom}}"/>
