--- src/components/images/img-comment/img-comment.scss
@@ -1,3 +1,0 @@
-img-comment {
-
-}
