+++ src/components/images/img-content/img-content.html
@@ -1,0 +1,1 @@
+<img *ngIf="srcRetail" [src]="srcRetail" [srcset]="srcSet" />
