+++ src/components/images/img-content/img-content.scss
@@ -1,0 +1,3 @@
+img-content {
+
+}
