+++ src/pages/articles-details/articles-details.html
@@ -14,7 +14,7 @@
         <ion-col>
           <ion-slides loop="{{articles.images.length > 1 ? true:false}}" autoplay="3500" pager="{{articles.images.length > 1 ? true:false}}">
             <ion-slide *ngFor="let image of articles.images">
-              <img-comment *ngIf="image.path" [src]=image.path></img-comment>
+              <img-content *ngIf="image.path" [src]=image.path></img-content>
               <div class="img-title">{{image.title}}</div>
             </ion-slide>
           </ion-slides>
