+++ src/pages/live/live.ts
@@ -21,4 +21,3 @@ export class LivePage {
   } 
   
 }
-
